pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "satlinkbudget"
dynamic = ["version"]
description = "Satellite link budget analysis library"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
]

[project.scripts]
satlinkbudget = "satlinkbudget.__main__:main"

[tool.hatch.version]
path = "src/satlinkbudget/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/satlinkbudget"]

[tool.pytest.ini_options]
testpaths = ["tests"]
